requirements_v2.txt
streamlit
google-genai
PyPDF2
python-pptx
pandas
openpyxl
requests
beautifulsoup4
fpdf2
pytesseract
Pillow